food-sage-main/food-sage-main/requirements.txt
flask==2.0.1
numpy==1.22.3
opencv-python==4.5.3.56
pillow==8.3.1
tensorflow==2.6.0